Move fallback card to a seperate file
